Complete protein-translation exercise.
Reflection:
I opted for a flat map for translating between codons and proteins.
Originally I wanted to use an array of codons for keys,
but I would've been forced to use an intermediary map to go from single
codons to an array grouping, then use that as a key into the final map.
The trade-off here is the duplicate values, and for a small exercise
it's worth it. I also opted to use for-of rather than
Array.prototype.map, as you cannot break execution in .map.