Complete space-age exercise
Reflection: While I was able to deduplicate some code with the private
class method, I'm not very happy with the extensibility of the class. If
you want to add a new planet, you have to create a separate method with
a new value. I'd rather the class be more data-driven and have the functions
be dynamically created at runtime.