Refactor space-age exercise to dynamically create functions
Reflection: This is my first attempt at dynamically creating functions
based on data. There are several trade-offs for this approach,
 the most significant one in my opinion is the lack of signatures for
the functions in my editor. Also I'm unsure of what the correct type
should be for the string index of the class. I can't use any public
members that aren't functions if I set the type to `Function`, and TS doesn't
like `Function | number`, since you can't call numbers like a function.
As a first attempt I'm happy with this approach, it just needs some
refining.